+++ lib/ezcater_rubocop.rb
@@ -19,6 +19,7 @@ RuboCop::ConfigLoader.instance_variable_set(:@default_configuration, config)
 require "rubocop/cop/ezcater/direct_env_check"
 require "rubocop/cop/ezcater/feature_flag_active"
 require "rubocop/cop/ezcater/feature_flag_name_valid"
+require "rubocop/cop/ezcater/graphql/not_authorized_scalar_field"
 require "rubocop/cop/ezcater/rails_configuration"
 require "rubocop/cop/ezcater/rails_env"
 require "rubocop/cop/ezcater/ruby_timeout"